refactor(server): split task executor pipeline
